Add AsMorphPivot trait
